+++ src/plugins/dam/index.js
@@ -5,6 +5,7 @@ import reducer from './reducers';
 import saga from './sagas';
 import { serviceIds } from './constants';
 import AssetSubscriber from './services/AssetSubscriber';
+import PatchAssetsSubscriber from './services/PatchAssetsSubscriber';
 import ImageAssetSubscriber from './services/ImageAssetSubscriber';
 
 export default class DamPlugin extends Plugin {
@@ -19,12 +20,14 @@ export default class DamPlugin extends Plugin {
 
     bottle.service(serviceIds.ASSET_SUBSCRIBER, AssetSubscriber);
     bottle.service(serviceIds.IMAGE_ASSET_SUBSCRIBER, ImageAssetSubscriber);
+    bottle.service(serviceIds.PATCH_ASSETS_SUBSCRIBER, PatchAssetsSubscriber);
   }
 
   getSubscriberServices() {
     return [
       serviceIds.ASSET_SUBSCRIBER,
       serviceIds.IMAGE_ASSET_SUBSCRIBER,
+      serviceIds.PATCH_ASSETS_SUBSCRIBER,
     ];
   }
 }